SaveOutlookMailsHere: Implementation of SaveMailsHere for Outlook .msg email files
Not yet tested on a real system with Microsoft Outlook installed!

But Drag&Drop from Outlook directly on this script icon does not work,
which is not a problem of this script, but a general drawback of the way
Drag&Drop is implemented in Outlook.

The "SenderEmailAddress" key can be set to a variable string for faking
different mails during test.