RenameOutlookMails: Variant of SaveOutlookMailsHere
Instead of the script directory, the directory of the email file is used
as destination, which effectively restricts the "move" operation to a
"rename".

Now it can be installed at one central place and be used in the Windows
Explorer's "Sent To" menu (reachable from context menu).

Only changed two lines of code.